Fail gracefully if parent repo doesn't exist or have specs (#67)
With the addition of #57, now rails grade will fail if the parent repo:
* Doesn't exist
* Doesn't have specs

Most of the time this happens when developing locally for a project that hasn't been published yet, but it's possible that it breaks if students fork a copy of a repo and name it differently.

This commit quits the task with a descriptive message if a project isn't available on appdev-projects or doesn't have a specs folder. This is to keep in line with the idea that the only specs that should be run are the ones available on appdev-projects.

+++ lib/tasks/grade.rake
@@ -133,21 +133,28 @@ def sync_specs_with_source
   reponame = `basename -s .git \`git config --get remote.origin.url\``.chomp
   full_reponame = "appdev-projects/#{reponame}"
 
-  repo_contents = Octokit.contents(full_reponame)
-  remote_spec_folder = repo_contents.find { |git_object| git_object[:name] == 'spec' }
-  remote_sha = remote_spec_folder[:sha]
-  # Discard unstaged changes in spec folder
-  `git checkout spec -q`
-  `git clean spec -f -q`
-  local_sha = `git ls-tree HEAD #{project_root.join('spec')}`.chomp.split[2]
+  if Octokit.repository?(full_reponame)
+    repo_contents = Octokit.contents(full_reponame)
+    remote_spec_folder = repo_contents.find { |git_object| git_object[:name] == 'spec' }
+    if remote_spec_folder.blank?
+      abort("The project #{full_reponame} does not have specs.")
+    end
+    remote_sha = remote_spec_folder[:sha]
+    # Discard unstaged changes in spec folder
+    `git checkout spec -q`
+    `git clean spec -f -q`
+    local_sha = `git ls-tree HEAD #{project_root.join('spec')}`.chomp.split[2]
 
-  unless remote_sha == local_sha
-    `git fetch upstream`
-    # Remove local contents of spec folder
-    `rm -rf spec/*`
-    default_branch = `git remote show upstream | grep 'HEAD branch' | cut -d' ' -f5`.chomp
-    # Overwrite local contents of spec folder with contents from upstream branch
-    `git checkout upstream/#{default_branch} spec/ -q`
+    unless remote_sha == local_sha
+      `git fetch upstream`
+      # Remove local contents of spec folder
+      `rm -rf spec/*`
+      default_branch = `git remote show upstream | grep 'HEAD branch' | cut -d' ' -f5`.chomp
+      # Overwrite local contents of spec folder with contents from upstream branch
+      `git checkout upstream/#{default_branch} spec/ -q`
+    end
+  else
+    abort("The project #{full_reponame} does not exist.")
   end
 end
